+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- AWS Region Endpoint for IRSA.
+
 ## [11.9.0] - 2022-04-20
 
 ### Added